+++ config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
+  resources :documents
   devise_for :users
   root 'home#index'
   get 'about', to: 'about#index'